update the NavBar.jsx/Missions.jsx and Rockets.jsx to persist the data entry for mission and rocket
