+++ src/components/App.js
@@ -1,11 +1,24 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { fetchMissions } from '../redux/missions/missionSlice';
+import { fetchRockets } from '../redux/Rockets/RocketsSlice';
 import NavBar from './NavBar';
 import Rockets from './Rockets';
 import Missions from './Missions';
 import MyProfile from './MyProfile';
 
 function App() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchMissions());
+  }, [dispatch]);
+
+  useEffect(() => {
+    dispatch(fetchRockets());
+  }, [dispatch]);
+
   return (
     <BrowserRouter>
       <NavBar />
